Renamed messages to backend-messages
